Extract code into separate function
